Add confirmation msg to enable/disable commands, restrict to authorised users

+++ process_dms.sh
@@ -24,7 +24,7 @@ is_existing_reply () {
 }
 
 send_not_authorized() {
-    send_reply "I'm sorry Dave, I'm afraid I can't do that."
+    send_dm_reply "I'm sorry Dave, I'm afraid I can't do that."
 }
 
 add_reply_string() {
@@ -33,7 +33,7 @@ add_reply_string() {
         # append string to replies
         NEW_REPLY=$(echo "$TWEETTEXT" | sed -e 's/^\+[[:space:]]*//')
         echo "$NEW_REPLY" >> $MYPATH/replies.txt
-        send_reply "New reply added: $NEW_REPLY"
+        send_dm_reply "New reply added: $NEW_REPLY"
     else
         send_not_authorized
     fi
@@ -47,21 +47,17 @@ delete_reply_string() {
 	if is_existing_reply "$TARGET_REPLY"
 	then
 	    sed -i "/$TARGET_REPLY/d" $MYPATH/replies.txt
-	    send_reply "Reply removed: $TARGET_REPLY"
+	    send_dm_reply "Reply removed: $TARGET_REPLY"
 	else
-            send_reply "I did not find that text in the current list of replies."
+            send_dm_reply "I did not find that text in the current list of replies."
         fi
     else
         send_not_authorized
     fi
 }
 
-send_reply() {
-    twidge -c $MYPATH/$CONFIG dmsend $SENDER "$1"
-}
-
 send_help_reply() {
-    send_reply "See https://github.com/cherdt/last-word-bot"
+    send_dm_reply "Commands: [ON|OFF|+ text|- text|URL]. See https://github.com/cherdt/last-word-bot"
 }
 
 get_tweet_info() {
@@ -71,25 +67,21 @@ get_tweet_info() {
     TWEETID=$(echo $TARGETTWEET | cut -d'/' -f 6)
 }
 
-get_random_reply() {
-    # Pick 1 reply at random
-    REPLY=$(shuf -n 1 $MYPATH/replies.txt)
-}
-
 # Message ID, sender, recipient, etc. are passed in by --exec
 ID=$1
 SENDER=$2
 RECIPIENT=$3
 TWEETTEXT=$4
 
-# Process command
-
-if [[ $TWEETTEXT =~ ^(ON|ENABLE)$ ]]
+# Process commands
+if [[ $TWEETTEXT =~ ^(ON|ENABLE)$ && is_authorized ]]
 then
     rm $MYPATH/.disabled
-elif [[ $TWEETTEXT =~ ^(OFF|DISABLE)$ ]]
+    send_on_confirmation
+elif [[ $TWEETTEXT =~ ^(OFF|DISABLE)$ is_authorized ]]
 then
     touch $MYPATH/.disabled
+    send_off_confirmation
 # if DM begins with "+" then we are adding a reply string
 elif [[ $TWEETTEXT =~ ^\+ ]]
 then
@@ -117,13 +109,13 @@ then
         # Reply to the message referenced by the DM
         twidge -c $MYPATH/$CONFIG update --inreplyto $TWEETID "@$TARGETUSER $REPLY"
     else
-        send_reply "I'm currently turned off. To turn me back on, an authorized user needs to send an ON command."
+        send_dm_reply "I'm currently turned off. To turn me back on, an authorized user needs to send an ON command."
     fi
 
 # otherwise, we didn't understand the command
 else
     logger "Failed to parse DM: $TWEETTEXT"
-    send_reply "Sorry, I didn't undertand that. Try HELP" 
+    send_dm_reply "Sorry, I didn't undertand that. Try HELP" 
 fi
 
 exit 0
